+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Added VeeValidate for form validation with custom rules
 - Added validated file input form component
 - Added tyre impression image upload modal
+- Added tyre model create modal and custom form elements
+- Added feather icons dependency
 
 ### [Changed]